scrub separately from index.js
